@@ -1,6 +1,6 @@
 module github.com/hashicorp/aws-sdk-go-base/v2
 
-go 1.20
+go 1.21
 
 require (
 	github.com/aws/aws-sdk-go-v2 v1.25.0
@@ -21,7 +21,6 @@ require (
 	github.com/mitchellh/go-homedir v1.1.0
 	go.opentelemetry.io/contrib/instrumentation/github.com/aws/aws-sdk-go-v2/otelaws v0.48.0
 	go.opentelemetry.io/otel v1.23.1
-	golang.org/x/exp v0.0.0-20230905200255-921286631fa9
 	golang.org/x/net v0.21.0
 	golang.org/x/text v0.14.0
 )
